Add files via upload
The following Project is a Multi Asset Portfolio made for a investor with Rs. 10Cr of investment to be invested in such a way that the investor get monthly cashflow of Rs. 3Lakh. The duration of investments are 5 Year and 10 Years i.e., the returns projection are made for both the duration. The monthly cashflows are pre tax calculated and inflation is not considered for the calculations.